# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Philips Hue Control API
  version: 2.0.0
  description: |2-

          A simplified, client-friendly API for controlling Philips Hue lights.

          **Features:**
          - Pre-computed RGB colors and CSS shadows
          - Unified dashboard endpoint (1 call instead of 4-6)
          - Room hierarchy already built
          - Session-based authentication
          - Demo mode for testing without a real bridge
          - Settings and weather API
          - Real-time updates via WebSocket

          **Authentication Methods:**
          1. **Session Token**: Get a session token from `POST /api/v1/auth/session`, then use `Authorization: Bearer <token>` header
          2. **Demo Mode**: Add `X-Demo-Mode: true` header to use mock data (no bridge required)

          **Demo Mode:**
          Try the API without a Hue Bridge by adding `X-Demo-Mode: true` header to any request.
          - Returns mock data (2 rooms, 6 lights, 4 scenes, 2 zones)
          - State changes persist in memory (reset on server restart)
          - WebSocket: authenticate with `{ type: 'auth', demoMode: true }`

          **Multi-Client Support:**
          When one client pairs with a bridge, credentials are stored server-side. Other clients can then connect without re-pairing:
          1. First client: `POST /api/v1/auth/pair` → `POST /api/v1/auth/session`
          2. Subsequent clients: `POST /api/v1/auth/connect` (no pairing needed!)
...

...

  contact:
...
  license:
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Session Token
      description: Session token from POST /api/v1/auth/session
    DemoMode:
      type: apiKey
      in: header
      name: X-Demo-Mode
      description: Set to "true" to use demo mode (no bridge required)
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: resource_not_found
        message:
          type: string
          example: The light 'abc123' was not found
        suggestion:
          type: string
          example: Check the light ID or refresh the dashboard to get current lights
...
    Location:
      type: object
      description: Geographic location for weather
      properties:
        lat:
          type: number
          example: 51.5074
          description: Latitude
        lon:
          type: number
          example: -0.1278
          description: Longitude
        name:
          type: string
          example: London
          description: Location name
      required:
        - lat
        - lon
    Settings:
      type: object
      description: User settings for the session
      properties:
        location:
          oneOf:
            - $ref: '#/components/schemas/Location'
            - type: 'null'
          description: Location for weather (null if not set)
        units:
          type: string
          enum:
            - celsius
            - fahrenheit
          example: celsius
          description: Temperature unit preference
        services:
          type: object
          description: Service enable/disable settings
          properties:
            hue:
              type: object
              properties:
                enabled:
                  type: boolean
                  description: Whether Hue service is enabled
            hive:
              type: object
              properties:
                enabled:
                  type: boolean
                  description: Whether Hive service is enabled
...
paths:
...
  /settings:
    get:
      summary: Get current settings
      description: Returns the current settings for this session (location and units)
      tags:
        - Settings
      security:
        - BearerAuth: []
        - DemoMode: []
      responses:
        '200':
          description: Current settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Settings'
        '401':
          description: Missing or invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update settings
      description: Update all settings (location and/or units)
      tags:
        - Settings
      security:
        - BearerAuth: []
        - DemoMode: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                location:
                  $ref: '#/components/schemas/Location'
                units:
                  type: string
                  enum:
                    - celsius
                    - fahrenheit
                services:
                  type: object
                  properties:
                    hue:
                      type: object
                      properties:
                        enabled:
                          type: boolean
                    hive:
                      type: object
                      properties:
                        enabled:
                          type: boolean
      responses:
        '200':
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Settings'
        '400':
          description: Invalid settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /settings/location:
    put:
      summary: Update location
      description: Update just the location setting
      tags:
        - Settings
      security:
        - BearerAuth: []
        - DemoMode: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Location'
      responses:
        '200':
          description: Location updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Settings'
        '400':
          description: Invalid location
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Clear location
      description: Remove the stored location
      tags:
        - Settings
      security:
        - BearerAuth: []
        - DemoMode: []
      responses:
        '200':
          description: Location cleared
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Settings'
  /settings/reset-demo:
    post:
      summary: Reset settings to demo defaults
      description: Resets all settings to demo defaults. Only available in demo mode. Used for E2E test isolation.
      tags:
        - Settings
      security:
        - DemoMode: []
      responses:
        '200':
          description: Settings reset successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Settings reset to demo defaults
        '403':
          description: Not in demo mode
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...